requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
Pillow==11.0.0
requests==2.32.3
python-multipart==0.0.16
pydantic==2.10.3